requirements.txt
ultralytics==8.1.2
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
matplotlib==3.7.2 